detalles

+++ app/Http/Controllers/UserController.php
@@ -12,6 +12,7 @@ use vendor\autoload;
 use \RouterOS\Query;
 use \RouterOS\Client;
 use App\DetailVoucher;
+use DB;
 use RealRashid\SweetAlert\Facades\Alert;
 
 class UserController extends Controller
@@ -77,8 +78,11 @@ public static function formatBytes($size, $precision = 2)
                     $total = $this->formatBytes($totalmemory);
                     $free = $this->formatBytes($freememory);
                     $rest = $this->formatBytes($resta);
+
+                    $getscripts = (new Query('/system/script/print'));
+                    $scripts = $client->query($getscripts)->read();
                 
-                    return view('mikvo.dashboard.layouts.main',["freememory"=>$free, "restmemeory"=>$rest,'costos'=>$costos,'usersall'=>$usersall, 'active'=>$active,'router'=>$inforouter, 'user' => $user ] );
+                    return view('mikvo.dashboard.layouts.main',["scripts"=>$scripts,"freememory"=>$free, "restmemeory"=>$rest,'costos'=>$costos,'usersall'=>$usersall, 'active'=>$active,'router'=>$inforouter, 'user' => $user ] );
                 }
                 
             }
@@ -88,7 +92,6 @@ public static function formatBytes($size, $precision = 2)
     }
     public function login(){
         if($this->isSession()){
-            Alert::info('Info Title', 'Info Message');
             return redirect(('/dashboard/routerboard'));
         }
         return view('welcome');
@@ -99,19 +102,19 @@ public static function formatBytes($size, $precision = 2)
         if(isset($usuario)){
             if(Hash::check($password_user,$usuario->password_user)){
                 session()->put('UserSession',$usuario);
-                Alert::toast('Bienvenido', 'success')->position('top-end')->autoClose(2000);
+                Alert::toast('Bienvenido', 'success')->position('top-end')->autoClose(3000);
                 //Alert::success('Success Title', 'Success Message')->autoClose(2000);
                 return redirect('/dashboard/routerboard');
             }       
         }
-        Alert::error('Error', 'revise su información')->autoClose(1000);
+        Alert::error('Error', 'revise su información')->autoClose(3000);
         return redirect('/login');
     }
 
     public function logout(){
             session()->forget('UserSession');
             session()->forget('routerConnected');
-            Alert::info('Sesión terminada','')->autoClose(1000);
+            Alert::info('Sesión terminada')->autoClose(3000);
             return redirect('/login');
         
     }
@@ -148,8 +151,10 @@ public static function formatBytes($size, $precision = 2)
                     $total = $this->formatBytes($totalmemory);
                     $free = $this->formatBytes($freememory);
                     $rest = $this->formatBytes($resta);
+                    $getscripts = (new Query('/system/script/print'));
+                    $scripts = $client->query($getscripts)->read();
                 
-                    return view('mikvo.dashboard.layouts.main',["freememory"=>$free, "restmemeory"=>$rest,'costos'=>$costos,'usersall'=>$usersall, 'active'=>$active,'router'=>$inforouter, 'user' => $user ] );
+                    return view('mikvo.dashboard.layouts.main',["scripts"=>$scripts,"freememory"=>$free, "restmemeory"=>$rest,'costos'=>$costos,'usersall'=>$usersall, 'active'=>$active,'router'=>$inforouter, 'user' => $user ] );
                 }
             }
             return redirect('/dashboard/routerboard');
@@ -160,20 +165,31 @@ public static function formatBytes($size, $precision = 2)
     public function register(Request $request){
         
         if(!session()->has('UserSession')){
-            $user = new User;
-            $user->fullname_user = $request->input('fullname_user');
-            $user->user_name = $request->input('user_name');
-            $user->telephone_user = $request->input('telephone_user');
-            $user->email_user = $request->input('email_user');
-            $user->password_user = Hash::make($request->input('password_user'));
+
+            $email = $request->input('email_user');
+            $emaildb = DB::table('users')->select('email_user')->where('email_user',$email)->first();
+            if($emaildb == null){
+                $user = new User;
+                $user->fullname_user = $request->input('fullname_user');
+                $user->user_name = $request->input('user_name');
+                $user->telephone_user = $request->input('telephone_user');
+                $user->email_user = $request->input('email_user');
+                $user->password_user = Hash::make($request->input('password_user'));
+                
+                $user->save();
+                Alert::success('Usuario registrado', 'Inicie sesión')->autoClose(1000);
+                return redirect('/login');
+            }else{
+                Alert::error('Ya existe un usuario con este correo', 'Verifica tu correo')->autoClose(3000);
+                return redirect('/register');
+            }
             
-            $user->save();
-            Alert::success('Usuario registrado', 'Inicie sesión')->autoClose(1000);
-        return redirect('/login');
+        }else{
+            $id = session()->get('UserSession')->id;
+            $user = User::find($id);
+            return view('mikvo.dashboard.layouts.main',['user'=>$user]);     
         }
-        $id = session()->get('UserSession')->id;
-        $user = User::find($id);
-        return view('mikvo.dashboard.layouts.main',['user'=>$user]);
+       
     }
 
 //------------------------------------------------------USER------------------------------------------------------    
@@ -200,23 +216,36 @@ public static function formatBytes($size, $precision = 2)
     {
         if(session()->has('UserSession')){
             $user = User::find($id);
-            $user->fullname_user = $request->input('fullname_user');
-            $user->user_name = $request->input('user_name');
-            $user->telephone_user = $request->input('telephone_user');
+            $email = $request->input('email_user');
 
-            if ($request->hasFile('img_user')) {
-                $imagen = explode(",", $user->img_user);
-                Storage::delete($user->img_user);
-                $user->img_user = $request->input('img_user');
-                $user->img_user = $request->file('img_user')->store('/');
+            $emaildb = DB::table('users')->select('email_user')->where('email_user',$email)->first();
+            if($emaildb == null){
+               
+                $user->email_user = $request->input('email_user');
+
+                $user->fullname_user = $request->input('fullname_user');
+                $user->user_name = $request->input('user_name');
+                $user->telephone_user = $request->input('telephone_user');
+                if ($request->hasFile('img_user')) {
+                    $imagen = explode(",", $user->img_user);
+                    Storage::delete($user->img_user);
+                    $user->img_user = $request->input('img_user');
+                    $user->img_user = $request->file('img_user')->store('/');
+                }
+                $user->save();
+                session()->forget('UserSession');
+                Alert::success('Datos actualizados','Inicia sesión nuevamente')->autoClose(3000);
+                return redirect('/login');
+            }else{
+                Alert::error('Ya existe un usuario con este correo')->autoClose(3000);
+                return redirect('/dashboard/user');
+
             }
-
-            $user->save();
-            return redirect('/dashboard/user')->with('message','Guardado Satisfactoriamente !');
+            
         }
-        return view('welcome');
+        
+       
     }
-
 
     //--------------------------------------------------UPDATE PASSWORD---------------------------------------------------    
 
@@ -239,9 +268,11 @@ public static function formatBytes($size, $precision = 2)
                 $user->password_user = Hash::make($password_repeat);
             }   
             $user->save();
-            return redirect('/dashboard/user')->with('message','Contraseña actualizada Satisfactoriamente !');            
+            Alert::success('Contraseña actualizada')->autoClose(3000);
+            return redirect('/dashboard/user');
         }else{
-            return redirect('/dashboard/user')->with('message','No se actualizo la contraseña!');
+            Alert::error('No se actualizo la contraseña')->autoClose(3000);
+            return redirect('/dashboard/user');
         }
     }
 }
